src/dremio/benchmark.py
import logging
import time
import json
from pathlib import Path
from typing import Dict, List, Any
import yaml
from concurrent.futures import ThreadPoolExecutor
import psutil
from datetime import datetime

from .client import DremioClient

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class DremioBenchmark:
    def __init__(self, config_path: str):
        """Initialize the benchmark runner."""
        self.config = self._load_config(config_path)
        self.source_client = DremioClient(config_path, "dremio_source")
        self.target_client = DremioClient(config_path, "dremio_target")
        self.results_dir = Path(self.config["benchmark"]["results"]["output_dir"])
        self.results_dir.mkdir(exist_ok=True)

    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """Load benchmark configuration."""
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)

    def prepare_environment(self, scale_factor: int) -> None:
        """Prepare the benchmark environment with test data."""
        logger.info(f"Preparing environment with {scale_factor}GB scale factor...")
        
        # Generate TPC-DS data
        self._generate_tpcds_data(scale_factor)
        
        # Convert to different formats
        for format in self.config["benchmark"]["data_preparation"]["file_formats"]:
            for compression in self.config["benchmark"]["data_preparation"]["compression"]:
                self._convert_data_format(format, compression)

    def _generate_tpcds_data(self, scale_factor: int) -> None:
        """Generate TPC-DS test data."""
        # Implementation for TPC-DS data generation
        # This would use dsdgen or similar tools
        pass

    def _convert_data_format(self, format: str, compression: str) -> None:
        """Convert data to specified format and compression."""
        # Implementation for format conversion
        pass

    def run_workload(self, workload_type: str) -> Dict[str, Any]:
        """Run a specific workload type."""
        logger.info(f"Running {workload_type} workload...")
        results = []
        
        for query in self.config["benchmark"]["workloads"][workload_type]:
            query_results = self._execute_query(query)
            results.append(query_results)
        
        return {
            "workload_type": workload_type,
            "queries": results,
            "timestamp": datetime.now().isoformat()
        }

    def _execute_query(self, query_config: Dict[str, Any]) -> Dict[str, Any]:
        """Execute a single query with monitoring."""
        query_name = query_config["name"]
        query = query_config["query"]
        
        # Warmup runs
        for i in range(self.config["benchmark"]["execution"]["warmup_runs"]):
            logger.info(f"Warmup run {i+1}/{self.config['benchmark']['execution']['warmup_runs']}")
            self.source_client.execute_query(query)
        
        # Measurement runs
        execution_times = []
        resource_usage = []
        
        for i in range(self.config["benchmark"]["execution"]["measurement_runs"]):
            logger.info(f"Measurement run {i+1}/{self.config['benchmark']['execution']['measurement_runs']}")
            
            # Start resource monitoring
            process = psutil.Process()
            start_cpu = process.cpu_percent()
            start_memory = process.memory_info().rss
            
            # Execute query
            start_time = time.time()
            result = self.source_client.execute_query(query)
            end_time = time.time()
            
            # End resource monitoring
            end_cpu = process.cpu_percent()
            end_memory = process.memory_info().rss
            
            execution_times.append(end_time - start_time)
            resource_usage.append({
                "cpu_percent": (start_cpu + end_cpu) / 2,
                "memory_bytes": end_memory - start_memory
            })
        
        return {
            "query_name": query_name,
            "execution_times": execution_times,
            "avg_time": sum(execution_times) / len(execution_times),
            "min_time": min(execution_times),
            "max_time": max(execution_times),
            "resource_usage": resource_usage,
            "rows_processed": len(result.get('rows', []))
        }

    def run_parallel_queries(self, queries: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """Run multiple queries in parallel."""
        with ThreadPoolExecutor(max_workers=self.config["benchmark"]["execution"]["parallel_queries"]) as executor:
            futures = [executor.submit(self._execute_query, query) for query in queries]
            return [future.result() for future in futures]

    def save_results(self, results: Dict[str, Any]) -> None:
        """Save benchmark results to file."""
        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
        
        # Save as JSON
        json_file = self.results_dir / f"benchmark_results_{timestamp}.json"
        with open(json_file, 'w') as f:
            json.dump(results, f, indent=2)
        
        # Save as CSV if configured
        if "csv" in self.config["benchmark"]["results"]["formats"]:
            self._save_results_csv(results, timestamp)
        
        logger.info(f"Results saved to {json_file}")

    def _save_results_csv(self, results: Dict[str, Any], timestamp: str) -> None:
        """Save results in CSV format."""
        import pandas as pd
        
        # Convert results to DataFrame
        rows = []
        for query in results["queries"]:
            for metric in self.config["benchmark"]["results"]["metrics"]:
                rows.append({
                    "timestamp": timestamp,
                    "query_name": query["query_name"],
                    "metric": metric,
                    "value": query.get(metric, None)
                })
        
        df = pd.DataFrame(rows)
        csv_file = self.results_dir / f"benchmark_results_{timestamp}.csv"
        df.to_csv(csv_file, index=False)
        logger.info(f"CSV results saved to {csv_file}")

    def run_all_benchmarks(self) -> Dict[str, Any]:
        """Run all configured benchmarks."""
        logger.info("Starting comprehensive benchmark suite...")
        
        all_results = {
            "timestamp": datetime.now().isoformat(),
            "objectives": [],
            "workloads": []
        }
        
        # Run each objective
        for objective in self.config["benchmark"]["objectives"]:
            objective_results = self._run_objective(objective)
            all_results["objectives"].append(objective_results)
        
        # Run each workload
        for workload_type in self.config["benchmark"]["workloads"]:
            workload_results = self.run_workload(workload_type)
            all_results["workloads"].append(workload_results)
        
        # Save results
        self.save_results(all_results)
        
        return all_results

    def _run_objective(self, objective: Dict[str, Any]) -> Dict[str, Any]:
        """Run a specific benchmark objective."""
        logger.info(f"Running objective: {objective['name']}")
        
        results = {
            "name": objective["name"],
            "description": objective["description"],
            "results": []
        }
        
        if objective["name"] == "query_performance":
            for size in objective["dataset_sizes"]:
                self.prepare_environment(size)
                size_results = self.run_workload("bi_reporting")
                results["results"].append({
                    "dataset_size_gb": size,
                    "performance": size_results
                })
        elif objective["name"] == "data_ingestion":
            # Run data ingestion benchmarks
            pass
        
        return results 