requirements.txt
requests>=2.25.0
pandas>=1.2.0
pyyaml>=5.4.0
psutil>=5.8.0
pytest>=6.0
pytest-cov>=2.0
